tidy up headers   - change .h to .hpp   - remove unneeded includes in .pro files   - add ostream_overloads header for use by both catch and spdlog to display Qt types   - add exprtk.hpp header to include it with desired options and without compiler warnings
