+++ ext/spdlog/tweakme.h
@@ -31,7 +31,7 @@
 //          You must set new pattern(spdlog::set_pattern(..") without any
 //          date/time in it
 //
-// #define SPDLOG_NO_DATETIME
+#define SPDLOG_NO_DATETIME
 ///////////////////////////////////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -41,7 +41,7 @@
 // WARNING: If the log pattern contains thread id (i.e, %t) while this flag is
 // on, the result is undefined.
 //
-// #define SPDLOG_NO_THREAD_ID
+#define SPDLOG_NO_THREAD_ID
 ///////////////////////////////////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -57,7 +57,7 @@
 // Uncomment if logger name logging is not needed.
 // This will prevent spdlog from copying the logger name on each log call.
 //
-// #define SPDLOG_NO_NAME
+#define SPDLOG_NO_NAME
 ///////////////////////////////////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////////////////////////////////
