+++ inc/geometry.hpp
@@ -14,8 +14,6 @@
 #include <QDebug>
 #include <QImage>
 
-#include "logger.h"
-
 namespace geometry {
 
 class Compartment {
