+++ test/src/main.cpp
@@ -1,5 +1,5 @@
 #define CATCH_CONFIG_RUNNER
-#include "catch_qt.h"
+#include "catch.hpp"
 
 #include <QApplication>
